requirements.txt
apex==0.9.10dev
attr==0.3.2
av==17.0.0
boto3==1.42.69
botocore==1.42.69
decord==0.6.0
deepspeed==0.18.8
demoji==1.1.0
docutils==0.22.4
dtw==1.4.0
editdistance==0.8.1
einops==0.8.2
fastdtw==0.3.4
ffmpeg==1.4
flash_attn==2.7.4.post1
ftfy==6.3.1
fvcore==0.1.5.post20221221
gym==0.26.2
h5py==3.16.0
habitat==0.3.3
habitat_baselines==0.3.320250127
habitat_sim==0.2.4.dev20230403
imageio==2.37.3
imgaug==0.4.0
importlib_metadata==8.7.1
iopath==0.1.10
ipdb==0.13.13
joblib==1.5.3
jsonlines==4.0.0
librosa==0.11.0
lmdb==1.8.1
matplotlib==3.10.8
mmcv==2.2.0
mmdet==3.3.0
msgpack_numpy==0.4.8
networkx==3.4.2
numpy==2.4.3
omegaconf==2.3.0
opencv_python==4.13.0.92
pandas==3.0.1
peft==0.18.1
Pillow==12.1.1
psutil==7.0.0
pyarrow==23.0.1
pycocoevalcap==1.2
pytorch_lightning==2.6.1
pytorch_transformers==1.2.0
pytorchvideo==0.1.5
PyYAML==6.0.2
PyYAML==6.0.3
regex==2026.2.28
Requests==2.32.5
sacred==0.8.7
safetensors==0.7.0
scikit_learn==1.8.0
scipy==1.17.1
setuptools==69.5.1
simplejson==3.20.2
six==1.17.0
skimage==0.0
tensorboardX==2.6.4
tensorflow==2.21.0
termcolor==3.3.0
timm==1.0.25
torch==2.7.0+cu128
torchaudio==2.7.0+cu128
torchmetrics==1.9.0
torchnet==0.0.4
torchvision==0.22.0+cu128
tqdm==4.66.2
transformers==5.2.0
tslearn==0.8.1
wandb==0.25.1
yacs==0.1.8
zsvision==0.7.12
